+++ skoolkit/skoolctl.py
@@ -37,6 +37,7 @@ AD_START = 'start'
 AD_ORG = 'org'
 AD_IGNOREUA = 'ignoreua'
 AD_LABEL = 'label'
+AD_REFS = 'refs'
 
 # An entry ASM directive is one that should be placed before the entry title
 # when it is associated with the first instruction in the entry